Update: coin graph 조회 옵션 추가 및 랭킹 조회 추가

+++ src/main/java/com/project/bumawiki/domain/coin/domain/Price.java
@@ -16,11 +16,11 @@ public class Price {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	private int price;
+	private Long price;
 
 	private LocalDateTime startedTime;
 
-	public Price(int price) {
+	public Price(Long price) {
 		this.price = price;
 		startedTime = LocalDateTime.now();
 	}
